Gradle-ify the project
