+++ build.gradle
@@ -1,0 +1,25 @@
+plugins {
+    id 'java'
+    id 'application'
+}
+
+repositories {
+    jcenter()
+    mavenCentral()
+}
+
+dependencies {
+    implementation group: 'com.google.code.gson', name:'gson', version: '2.8.6'
+    implementation group: 'net.lingala.zip4j', name: 'zip4j', version: '2.2.3'
+
+    testImplementation group: 'org.junit.jupiter', name:'junit-jupiter-api', version: '5.4.2'
+    testRuntimeOnly group: 'org.junit.jupiter', name:'junit-jupiter-engine', version: '5.4.2'
+}
+
+application {
+    mainClassName = 'vazkii.cmpdl.CMPDL'
+}
+
+test {
+    useJUnitPlatform()
+}
